internal/civisibility: refactor
